Backend/requirements.txt
fastapi==0.110.0
pymongo==4.6.3
motor==3.4.0
python-dotenv==1.0.1
uvicorn[standard]
passlib[bcrypt]
PyJWT==2.8.0
httpx==0.26.0